put Cell and Board code in separate files for enhanced readibility
