+++ sudoku.py
@@ -1,192 +1,7 @@
 
 import pygame
 import random 
-from sudoku_generator import SudokuGenerator
-
-
-
-
-class Cell:
-
-    def __init__(self, value, row, col, screen):
-
-        self.value = value
-
-        self.row = row
-
-        self.col = col
-
-        self.screen = screen
-
-
-    def set_cell_value(self, value):
-
-        if not self.filled:
-
-            self.value = value
-
-
-    def set_sketched_value(self, value):
-
-        self.sketched_value = value 
-
-
-    def draw(self, surface):
-
-        rect = pygame.Rect(self.value, self.row, self.col)
-
-        pygame.draw.rect(surface, (255, 255, 255), rect)
-
-        border_color = (255,0,0)
-
-        pygame.draw.rect(surface, border_color, rect, 2)
-
-
-        if self.value != 0:
-
-            font = pygame.font.Font(None, self.size - 10)
-
-            text = font.render(str(self.value), True, (0,0,0))
-
-            text_rect = text.get_rect(center = rect.center)
-
-            surface.blit((text, text_rect))
-
-
-class Board:
-
-    def __init__(self, width, height, screen, difficulty):
-
-        self.width = width
-
-        self.height = height 
-
-        self.screen = screen
-
-        self.difficulty = difficulty 
-
-        self.cells = []
-
-        self.selected_row = None
-
-        self.selected_col = None
-
-        removed_squares = 0 
-
-        if difficulty == "easy":
-
-            removed_squares = 10
-
-        elif difficulty == "medium":
-
-            removed_squares = 20
-
-        else:
-            
-            removed_squares = 30 
-
-        sudoku = SudokuGenerator(self.width, removed_squares)
-
-        sudoku.fill_values()
-
-
-   
-
-
-        sudoku.remove_cells()
-
-        my_Board = sudoku.get_board()
-
-
-        for row in range(self.width):
-
-            for col in range(self.height):
-
-                cell_value = my_Board[row][col] 
-
-                self.cells[row][col] = Cell(cell_value, row, col, screen)
-
-
-        def draw(self):
-
-            pass
-
-
-        def select(self, row, col):
-            
-            self.select_row = row
-
-            self.select_col = col
-
-
-
-            
-
-        def click(self, x, y):
-
-            pass
-
-        def clear(self):
-
-            pass
-
-        def sketch(self, value):
-
-            pass
-
-
-        def place_number(self, value):
-
-            if self.selected_row is not None and self.selected_col is not None:
-
-                my_Board[self.selected_row][self.selected_col] = value
-
-                
-
-        def reset_to_original(self):
-
-            pass
-
-        def is_full(self):
-
-            for row in range(self.width):
-
-                for col in range(self.height):
-
-                    if my_Board[col][row] == "0":
-
-                        return False
-                    
-            else:
-
-                return True 
-
-        def update_board(self):
-
-            pass
-
-        def find_empty(self):
-
-            for row in range(self.width):
-
-                for col in range(self.height):
-
-                    if my_Board[col][row] == "0":
-
-                        return (col, row)
-
-        def check_board(self):
-
-            pass
-    
-
-   
-
-
-        
-
-
-
+from sudoku_generator import * 
 
 def main():
     try:
